[#78] Moved rebin functional tests to separate file
